# Implement write_map_to as serializing a SourceMap struct