This is a first pass to see what an implementation that no longer requires the disclaimer "TODO this is borderline unmaintainable" would look like. I have not profiled this other than running the existing mini benchmarks, where the times are identical to the previous implementation. Do you already have a more substantial benchmark that you have been consulting locally?