@dtolnay 2018-07-20
I measured a modest performance improvement from this PR against the `write_map_to` benchmark in #46.

- Before: `32,958,172 ns/iter (+/- 1,019,759) = 451 MB/s`
- After: `30,034,932 ns/iter (+/- 1,659,953) = 495 MB/s`