@dtolnay 2018-07-20
Rebased to fix conflict with 797ce7c864f6e3badbbb0091d51abf41e0b9498f.